[Globals] Move Poster struct into separate file
