+++ src/globals/Poster.cpp
@@ -1,0 +1,3 @@
+#include "globals/Poster.h"
+
+// no-op
